+++ applications/services/net/eth_net/eth_net.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "mongoose.h"
 #include "net_utils.h"
 #include "net.h"